feat(newsletterForm): add using of the newsletter component
