+++ landing-pages/src/components/navigation/footer/NewsletterSection.tsx
@@ -1,65 +1,19 @@
-import { useState } from 'react';
-import { Box, Button, Input, Stack, Text, useToast, VStack } from '@chakra-ui/react';
-import { Link } from '@chakra-ui/next-js';
+import { Text, VStack } from '@chakra-ui/react';
 
-import subscribeToNewsletter from 'utils/subscribeToNewsletter';
+import NewsletterForm from 'components/NewsletterForm';
 
-const NewsletterSection = (): JSX.Element => {
-	const [newsletterEmail, setNewsletterEmail] = useState('');
-
-	const toast = useToast({ duration: 3000, isClosable: true });
-
-	return (
-		<VStack align="start" maxW="400px" spacing="16px">
-			<VStack w="100%" align="start" spacing="0px">
-				<Text size={{ base: 'xl', xl: '2xl' }} color="white" id="edgar-footer-newsletterTitle-text">
-					SUIVEZ NOS ACTUALITÉS !!
-				</Text>
-				<Text color="white" id="edgar-footer-newsletterSubtitle-text">
-					Inscrivez-vous et profitez d'un accès en exclusivité lors de l'ouverture de notre beta
-				</Text>
-			</VStack>
-			<VStack w="100%" align="start" spacing="16px">
-				<Input
-					type="email"
-					w="100%"
-					value={newsletterEmail}
-					placeholder="prenom.nom@gmail.com"
-					textColor="white"
-					border="2px solid"
-					borderColor="green.500"
-					_placeholder={{
-						color: 'gray.300',
-					}}
-					_hover={{
-						borderColor: 'green.500',
-					}}
-					onChange={(e) => setNewsletterEmail(e.target.value)}
-					id="edgar-footer-newsletterEmail-input"
-				/>
-				<Stack direction={{ base: 'column', xl: 'row' }} justify="space-between" w="100%">
-					<Text color="white" maxW="240px" id="edgar-footer-newsletterPolicies-text">
-						En vous abonnant, vous acceptez notre{' '}
-						<Link href="">
-							<Box as="u">politique de confidentialité</Box>
-						</Link>
-					</Text>
-					<Button
-						variant="primaryBordered"
-						onClick={() => {
-							subscribeToNewsletter(newsletterEmail).then((res) => {
-								toast(res);
-								if (res.status !== 'error') setNewsletterEmail('');
-							});
-						}}
-						id="edgar-footer-newsletterSubscription-button"
-					>
-						M'abonner
-					</Button>
-				</Stack>
-			</VStack>
+const NewsletterSection = (): JSX.Element => (
+	<VStack align="start" maxW="400px" spacing="16px">
+		<VStack w="100%" align="start" spacing="0px">
+			<Text size={{ base: 'xl', xl: '2xl' }} color="white" id="edgar-footer-newsletterTitle-text">
+				SUIVEZ NOS ACTUALITÉS !!
+			</Text>
+			<Text color="white" id="edgar-footer-newsletterSubtitle-text">
+				Inscrivez-vous et profitez d'un accès en exclusivité lors de l'ouverture de notre beta
+			</Text>
 		</VStack>
-	);
-};
+		<NewsletterForm variant="blue" id="footer" />
+	</VStack>
+);
 
 export default NewsletterSection;
